- Created QueryParameters superclass to remove code duplication - Fixed typo
